test: refactor existing backfill tests
